requirements.txt
awsebcli==3.7.7
Babel==2.3.4
blessed==1.9.5
boto==2.42.0
botocore==1.4.41
cement==2.8.2
cffi==1.7.0
colorama==0.3.7
cryptography==1.4
defusedxml==0.4.1
Django==1.8.14
django-appconf==1.0.2
django-compressor==2.0
django-extra-views==0.6.4
django-haystack==2.4.1
django-localflavor==1.3
django-mobile==0.7.0
django-oscar==1.2.2
django-oscar-paypal==0.9.7
django-storages==1.5.0
django-tables2==1.0.7
django-treebeard==4.0.1
django-widget-tweaks==1.4.1
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
factory-boy==2.6.1
fake-factory==0.5.9
idna==2.1
jmespath==0.9.0
mock==1.3.0
mysqlclient==1.3.7
oauthlib==1.1.2
pathspec==0.3.4
paypalrestsdk==1.11.7
pbr==1.10.0
phone==0.4.0
phonenumbers==7.5.1
Pillow==3.3.0
purl==1.3
pyasn1==0.1.9
pycparser==2.14
PyJWT==1.4.1
pyOpenSSL==16.0.0
python-dateutil==2.5.3
python-social-auth==0.2.19
python3-openid==3.0.10
pytz==2016.6.1
PyYAML==3.11
rcssmin==1.0.6
requests==2.9.1
requests-oauthlib==0.6.2
rjsmin==1.0.12
s3transfer==0.1.3
semantic-version==2.5.0
six==1.10.0
sorl-thumbnail==12.3
texttable==0.8.4
Unidecode==0.4.19
wcwidth==0.1.7
websocket-client==0.37.0
